requirements/local.txt
-r ./base.txt

flake8==3.6.0
coverage==4.5.2

django-debug-toolbar==1.11
django-coverage-plugin==1.6.0
